requirements.txt
alembic==1.0.10
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.4.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.0
urllib3==1.24.3
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
